New branch for development of the Unit system
Creation of playable unit structures
Branch Goals / Requirements
1.	Units are able to deal and receive damage
2.	Units can move across the battle map
3.	Units have a defined movement cost matrix. This is later used in the pathfinding phase.
4.	Units can be defeated
5.	If time allows, create items that units can equip to modify their attributes

+++ Source/MMTSGFramework/Private/BaseUnit.cpp
@@ -32,9 +32,9 @@ void ABaseUnit::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent)
 
 }
 
-int64 ABaseUnit::GetMovementCost(ETileTerrainType)
+int32 ABaseUnit::GetMovementCost(ETileTerrainType)
 {
-	int64 MovementCost = 1;
+	int32 MovementCost = 1;
 
 	return MovementCost;
 }
